Add new API function ihex_byte_copy().
The implementation is basically the same as the old ihex_mem_copy()
function.  However, the destination memory is not zeroed to avoid
overwriting content which is not specified in the ihex records.

The copy is done completely byte-wise, avoiding memory access
violations.  Record contents whose address lies beyond the given
destination range are silently discarded instead of aborting with an
IHEX_ERR_ADDRESS_OUT_OF_RANGE error code.